Beginning Steps
So far, I have written the code for the m-statistic and d-statistic
within the function. I’m still figuring out what will be inside the
argument of the function and how to make the input include vectors and
matrices. I’m also still practicing and understanding the process
behind functions, let alone how to get the correct output for this
specific problem. For the time being, I put a “…” inside the
parentheses as a placeholder that I will fill in later.