Setting Up the Input to Allow for a Matrix and a Vector
As I specified in my comments, I am trying to get my Benford’s Law
function to take a hypothetical example of a random sample of vote
counts across three candidates in 20 counties. For the actual function
itself, I want it to accept the proportions of the first digits,
meaning that if the first digit “1” occurs 30% of the time, I want the
function to reflect that by inputting .3. I attempted to do that with
the proportion_matrix, but I am thinking that I will take that part
out. I’m doing my best to figure this out and I might simplify the
hypothetical example going forward to make the code simpler.