Completed Problem Set 2
I completed the problem set to the best of my ability and at this
point, I am concerned about getting the process correct. To my
knowledge, the first function does not return an output despite R
claiming that no errors occurred when I wrote the code for the
function. Naturally, this affects the output for the second function.
I’ll be asking about this process next week and I want to ensure that I
can continue to grow in my coding knowledge. I absolutely want to get
to the point where I can both get the process and output correct and
I’ll continue to work as much as I can to improve.